Suggestion for syntax improvement
To enable flexible call outs on a variety of platforms, a very small syntax change could help and also enable value added platforms consuming this data in the future. If we capture the platform the handles are on, we increase the discoverability of the oneforones. Plus, we can then create code that just applies the handle to the URL of the platform to go to the direct link (oneoronehandle + oneforoneplatform).